backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
pillow==10.1.0
requests==2.31.0
openai==1.3.7
httpx==0.25.2
python-multipart==0.0.6
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0
pydantic==2.5.0
aiofiles==23.2.1